Added Admin feature

+++ src/server/domains/products/router.ts
@@ -1,5 +1,6 @@
 import Router from 'express';
 import ProductService from './service';
+import { verifyToken } from '../../middleware/authenticate';
 
 export default function ProductRouter(productService: ProductService) {
     const router = Router();
@@ -25,8 +26,11 @@ export default function ProductRouter(productService: ProductService) {
         }
     });
 
-    router.post('/', async (req, res) => {
+    router.post('/', verifyToken, async (req, res) => {
         try {
+            if (req.body.credential.roles.indexOf('admin') === -1) {
+                return res.status(403).send({message: 'You do not have permission to register products'});
+            } 
             const product = await productService.createProduct(req.body);
             res.status(201).send(product);
         } catch (error) {
@@ -34,8 +38,11 @@ export default function ProductRouter(productService: ProductService) {
         }
     });
 
-    router.put('/:id', async (req, res) => {
+    router.put('/:id', verifyToken, async (req, res) => {
         try {
+            if (req.body.credential.roles.indexOf('admin') === -1) {
+                return res.status(403).send({message: 'You do not have permission to update products'});
+            } 
             const product = await productService.updateProduct(req.params.id, req.body);
             if (!product) {
                 return res.status(404).send({ message: 'Product not found' });
@@ -46,8 +53,11 @@ export default function ProductRouter(productService: ProductService) {
         }
     });
 
-    router.delete('/:id', async (req, res) => {
+    router.delete('/:id', verifyToken, async (req, res) => {
         try {
+            if (req.body.credential.roles.indexOf('admin') === -1) {
+                return res.status(403).send({message: 'You do not have permission to delete products'});
+            } 
             const product = await productService.deleteProduct(req.params.id);
             if (!product) {
                 return res.status(404).send({ message: 'Product not found' });
